backend/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
APScheduler==3.11.0
arrow==1.3.0
async-timeout==5.0.1
attrs==25.3.0
bcrypt==4.0.1
boolean.py==5.0
CacheControl==0.14.3
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
coverage==7.8.2
cyclonedx-python-lib==9.1.0
defusedxml==0.7.1
ecdsa==0.19.1
exceptiongroup==1.3.0
fastapi==0.115.11
fastapi-cli==0.0.7
filelock==3.18.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
hypothesis==6.135.20
idna==3.10
iniconfig==2.1.0
license-expression==30.4.3
markdown-it-py==3.0.0
mdurl==0.1.2
msgpack==1.1.1
packageurl-python==0.17.1
packaging==25.0
passlib==1.7.4
pip-api==0.0.34
pip-requirements-parser==32.0.1
pip_audit==2.9.0
platformdirs==4.3.8
pluggy==1.6.0
psutil==7.0.0
psycopg2-binary==2.9.10
py-serializable==2.0.0
pyasn1==0.4.8
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.2
pyparsing==3.2.3
pytest==8.3.5
pytest-cov==6.1.1
pytest-subtests==0.14.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
redis==6.2.0
requests==2.32.3
rich==14.0.0
rich-toolkit==0.14.8
rsa==4.9
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.38
starlette==0.46.1
starlette-testclient==0.4.1
toml==0.10.2
tomli==2.2.1
typer==0.15.2
types-python-dateutil==2.9.0.20250516
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.3.0
uvicorn==0.34.0
watchfiles==1.1.0
websockets==15.0.1
